Remove migration files and controllers
The commit includes the removal of various migration files associated with the creation and modification of activity log tables. Additionally, it also removed several controller files. These changes seem to be associated with cleanup or refactoring as part of a potential change in the software architecture or removal of unused features.